backend/requirements.txt
# Required Python packages for the backend of the Honeypot Detector Pro
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
etherscan-python==2.1.0
colorama==0.4.6
python-dotenv==1.0.0